+++ go/understackctl/cmd/argocd/templates/argoCluster.tmpl
