+++ go/understackctl/cmd/node/node.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"os"
 
-	"github.com/rackerlabs/understack/go/understack/helpers"
+	"github.com/rackerlabs/understack/go/understackctl/helpers"
 
 	"github.com/charmbracelet/log"
