+++ go/understackctl/cmd/openstack/secrets.go
@@ -1,7 +1,7 @@
 package openstack
 
 import (
-	"github.com/rackerlabs/understack/go/understack/helpers"
+	"github.com/rackerlabs/understack/go/understackctl/helpers"
 
 	"github.com/charmbracelet/log"
