+++ go/understackctl/cmd/quickstart/quickstart.go
@@ -7,13 +7,13 @@ import (
 
 	"github.com/charmbracelet/log"
 	"github.com/gookit/goutil/envutil"
-	"github.com/rackerlabs/understack/go/understack/cmd/argocd"
-	"github.com/rackerlabs/understack/go/understack/cmd/certManager"
-	"github.com/rackerlabs/understack/go/understack/cmd/dex"
-	"github.com/rackerlabs/understack/go/understack/cmd/helmConfig"
-	"github.com/rackerlabs/understack/go/understack/cmd/node"
-	"github.com/rackerlabs/understack/go/understack/cmd/openstack"
-	"github.com/rackerlabs/understack/go/understack/cmd/other"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/argocd"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/certManager"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/dex"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/helmConfig"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/node"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/openstack"
+	"github.com/rackerlabs/understack/go/understackctl/cmd/other"
 	"github.com/spf13/cobra"
 )
