+++ go/understackctl/helpers/random.go
